pyproject.toml
[project]
name = "trainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "accelerate>=1.13.0",
    "bitsandbytes>=0.49.2",
    "datasets>=4.8.4",
    "ninja>=1.13.0",
    "packaging>=26.0",
    "peft>=0.18.1",
    "pyyaml>=6.0.3",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.5.0",
    "trl>=1.0.0",
    "wandb>=0.25.1",
]


[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
torchaudio = { index = "pytorch-cu128" }
